Extract camera to object; Refactor renderer;
